Normalizing newlines given to ensure correct behavior regardless of code formatting as checked into source control.
This is an issue in some shell environments not coping with windows style newlines which may creep into source control. Errors such as `parse error near '\n'` can occur when dot sourcing the profile.
Since we are injecting text into platform specific files, it is necessary to ensure that line endings are normalized as late as possible (i.e. at runtime).