Specify which options can be specified multiple times
Add details to our json help output so that we know which options can be specified multiple times. ie, for the `deploy-release` command, you can specify the `--skip` parameter multiple times, once for each step you want to skip.

This is useful information for external tools (ie nuke.build) to know when they are code generating wrappers for octopus cli.