Test tentacle separately, with soft failure for now (#75)
Because the primary purpose of this script is to test Octopus CLI, we can test that separately, and first.

The quick tentacle smoke test is a secondary thing that is currently not supported yet on ubuntu focal, so this change also makes that a soft failure for now. I think it's still useful to leave in the process for diagnostic reasons.